+++ src/main/java/com/project/wineshop/TypeNotFoundException.java
@@ -2,7 +2,7 @@ package com.project.wineshop;
 
 public class TypeNotFoundException extends RuntimeException {
 
-    TypeNotFoundException(Long id) {
+    TypeNotFoundException(Integer id) {
         super("Could not find type " + id);
     }
 }
