+++ src/main/java/com/project/wineshop/TypeRepository.java
@@ -2,6 +2,6 @@ package com.project.wineshop;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-interface TypeRepository extends JpaRepository<Type, Long> {
+interface TypeRepository extends JpaRepository<Type, Integer> {
     Type findByName(String name);
 }
